Re-adds multiple auth structure it seems like it is best practice
